Make null processing more consistent
If there was at least one value for a given dimension in a segment, then null values would be returned for all of the rows that the value didn't appear on.  But if you then had a segment that doesn't have any values for the given dimension, it would effectively filter out all of those rows and not aggregate them together.  Make it more consistent such that if you query for a dimension that doesn't exist, it returns null values.  Querying for a dimension that cannot be used as a dimension (like a metric) will also generate null values.